requirements.txt
pymisp
pyyaml
requests
rich
requests